Add Person super class and refactor student, teacher and apprentice_teacher classes
